+++ src/networkcloud/azext_networkcloud/aaz/latest/networkcloud/virtualmachine/_list.py
@@ -26,10 +26,10 @@ class List(AAZCommand):
     """
 
     _aaz_info = {
-        "version": "2022-12-12-preview",
+        "version": "2023-05-01-preview",
         "resources": [
-            ["mgmt-plane", "/subscriptions/{}/providers/microsoft.networkcloud/virtualmachines", "2022-12-12-preview"],
-            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.networkcloud/virtualmachines", "2022-12-12-preview"],
+            ["mgmt-plane", "/subscriptions/{}/providers/microsoft.networkcloud/virtualmachines", "2023-05-01-preview"],
+            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.networkcloud/virtualmachines", "2023-05-01-preview"],
         ]
     }
 
@@ -118,7 +118,7 @@ class List(AAZCommand):
         def query_parameters(self):
             parameters = {
                 **self.serialize_query_param(
-                    "api-version", "2022-12-12-preview",
+                    "api-version", "2023-05-01-preview",
                     required=True,
                 ),
             }
@@ -197,6 +197,10 @@ class List(AAZCommand):
             properties.admin_username = AAZStrType(
                 serialized_name="adminUsername",
                 flags={"required": True},
+            )
+            properties.availability_zone = AAZStrType(
+                serialized_name="availabilityZone",
+                flags={"read_only": True},
             )
             properties.bare_metal_machine_id = AAZStrType(
                 serialized_name="bareMetalMachineId",
@@ -464,7 +468,7 @@ class List(AAZCommand):
         def query_parameters(self):
             parameters = {
                 **self.serialize_query_param(
-                    "api-version", "2022-12-12-preview",
+                    "api-version", "2023-05-01-preview",
                     required=True,
                 ),
             }
@@ -543,6 +547,10 @@ class List(AAZCommand):
             properties.admin_username = AAZStrType(
                 serialized_name="adminUsername",
                 flags={"required": True},
+            )
+            properties.availability_zone = AAZStrType(
+                serialized_name="availabilityZone",
+                flags={"read_only": True},
             )
             properties.bare_metal_machine_id = AAZStrType(
                 serialized_name="bareMetalMachineId",