+++ src/containerapp/HISTORY.rst
@@ -5,6 +5,10 @@ Release History
 
 Upcoming
 ++++++
+* 'az containerapp create/update': --termination-grace-period support custom termination grace period
+
+0.3.34
+++++++
 * 'az containerapp job execution show/list': improve table output format
 * 'az containerapp create/update': --yaml support properties for api-version 2023-04-01-preview (e.g. subPath, mountOptions)
 * 'az containerapp service': add support for creation and deletion of kafka
@@ -13,6 +17,7 @@ Upcoming
 * Add regex to fix validation for containerapp name
 * Add 'az containerapp ingress cors' for CORS support
 * 'az container app env create/update': support --enable-mtls parameter
+* 'az containerapp up': fix issue where --repo throws KeyError
 
 0.3.33
 ++++++